Move external-tool related code into "pwiz_tools\Skyline\Controls"
