+++ pwiz_tools/Skyline/Alerts/LocateFileDlg.cs
@@ -19,9 +19,9 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using System.Windows.Forms;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Util;
 using pwiz.Skyline.Util.Extensions;
