+++ pwiz_tools/Skyline/Model/DdaSearch/MsFraggerSearchEngine.cs
@@ -21,7 +21,6 @@ using pwiz.Common.Collections;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Model.DocSettings;
 using pwiz.Skyline.Model.Results;
-using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
 using System;
@@ -34,6 +33,7 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
+using pwiz.Skyline.Controls.Tools;
 
 namespace pwiz.Skyline.Model.DdaSearch
 {
