+++ pwiz_tools/Skyline/Program.cs
@@ -36,8 +36,8 @@ using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
 using pwiz.Skyline.Controls.Startup;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model;
-using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
 using pwiz.Skyline.Util.Extensions;
