+++ pwiz_tools/Skyline/TestFunctional/ToolStoreDlgTest.cs
@@ -23,8 +23,8 @@ using System.IO;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline.Alerts;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Model.DocSettings;
-using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.ToolsUI;
 using pwiz.Skyline.Util;
