+++ pwiz_tools/Skyline/ToolsUI/RInstaller.cs
@@ -28,7 +28,7 @@ using System.Windows.Forms;
 using Microsoft.Win32;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
-using pwiz.Skyline.Model.Tools;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
 using pwiz.Skyline.Util.Extensions;
