+++ pwiz_tools/Skyline/ToolsUI/ToolStoreDlg.cs
@@ -30,7 +30,7 @@ using JetBrains.Annotations;
 using Newtonsoft.Json;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
-using pwiz.Skyline.Model.Tools;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
