+++ pwiz_tools/Skyline/Util/UtilInstall.cs
@@ -24,7 +24,7 @@ using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using Ionic.Zip;
-using pwiz.Skyline.Model.Tools;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.Properties;
 
 namespace pwiz.Skyline.Util
