+++ pwiz_tools/Skyline/Util/UtilUI.cs
@@ -27,9 +27,9 @@ using System.Threading;
 using System.Windows.Forms;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Alerts;
+using pwiz.Skyline.Controls.Tools;
 using pwiz.Skyline.EditUI;
 using pwiz.Skyline.Model;
-using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using ZedGraph;
